\name{optimlag-class}
\docType{class}
\alias{optimlag-class}
\title{
Representation of class \code{optimlag}
}

\description{
This class contains several useful objects about the optimum lag order for further manipulations.
}


\arguments{
  \item{Olag.value}{Under the assumption of only one common trend, the lag order to maximize Johansen Statistic(1988).
}
\item{Olag}{ The mazimized corresponding Johansen Statistic. 
}
\item{list.lags}{ A list of all the possible lag orders and their corresponding Johansen Statistics.
}

}


\author{
Fan Yang
}

\keyword{ optimlag }
