\name{print.ComT}
\alias{print.ComT}
\title{ Method for printing the class \code{ComT} }
\description{
  Displays the outcome of the \code{Kasa.ComT} and \code{GG.ComT}.
}

\seealso{
  \code{\link{ComT-class}}, \code{\link{Kasa.ComT}},
  \code{\link{GG.ComT}}.
}

\author{Fan Yang}
\keyword{methods}
